This program is suppose to send the commands PM 1\r BM 1\r GH\r
to the cmucam2 and compares it with some values I got from the stop sign using the cmucam and sees if it matches close to.  Hopefully it'll recognize if what is in front of it is a stop sign or not.

Haven't tested it on the actual arduino, but if anyone gots like a few minutes, test it for me and just paste me the output, that would be great.